fixes mocking of registry functions
